Stop pretending to run RegressionTests on main
RegressionTests is comparing main's HEAD to itself on pushes. Previously I let this slide because I appreciated stress testing RegressionTests for false positives. RegressionTests has not reported a false positive comparing a commit to itself in ages so it's time to stop wasting CI on this.

In the future, perhaps we can add these back and make them compare to the previous commit.